% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciFbTwo.rho_v2.R
\name{ciFbTwo.rho}
\alias{ciFbTwo.rho}
\title{Confidence Interval for F.beta Difference}
\usage{
ciFbTwo.rho(
  data,
  rho.N = NULL,
  rho.P = NULL,
  beta = 1,
  alpha = 0.05,
  normal.approx = c("auto", "TRUE", "FALSE")
)
}
\arguments{
\item{data}{A data frame (columns: actual outcomes,
C1 predicted outcomes, C2 predicted outcomes).
Encoded as 0 (negative) and 1 (positive).}

\item{rho.N}{Correlation of two classifiers' predictions
among negative cases. rho.N will be estimated if rho.N = NULL.}

\item{rho.P}{Correlation of two classifiers' predictions
among positive cases. rho.P will be estimated if rho.P = NULL.}

\item{beta}{beta.}

\item{alpha}{Significance level.}

\item{normal.approx}{normal.approx = c("auto", "TRUE", "FALSE").}
}
\description{
A function to calculate the CI of the F.beta difference between
two correlated classifiers.
}
\examples{
\dontrun{
set.seed(2028)
data <- data_Generation(N=100, s=40,
                        pp1=0.6, ps1=0.8,
                        pp2=0.7, ps2=0.9,
                        rho.N=0.2, rho.P=0.2,
                        beta=1)
ciFbTwo.rho(data)
ciFbTwo.rho(data, rho.N=0, rho.P=0) # wider
}
}
