% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{data_Generation}
\alias{data_Generation}
\title{Data Generation}
\usage{
data_Generation(
  N = 100,
  s = 40,
  pp1 = 0.6,
  ps1 = 0.8,
  pp2 = 0.7,
  ps2 = 0.9,
  rho.N = 0,
  rho.P = 0,
  beta = 1
)
}
\arguments{
\item{N}{Number of classification instances.}

\item{s}{Number of positives.}

\item{pp1}{Precision of 1st classifier.}

\item{ps1}{Sensitivity of 1st classifier.}

\item{pp2}{Precision of 2nd classifier.}

\item{ps2}{Sensitivity of 2nd classifier.}

\item{rho.N}{Correlation of two classifiers' predictions
among negative cases.}

\item{rho.P}{Correlation of two classifiers' predictions
among positive cases.}

\item{beta}{beta.}
}
\description{
A function to generate the predicted results of two classifiers.
}
