% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFb.diff.rho_cpp.R
\name{F1.cond.two.b.rho}
\alias{F1.cond.two.b.rho}
\title{Distribution of the Difference between Two F Beta Scores}
\usage{
F1.cond.two.b.rho(
  N,
  s,
  pp1 = 9/11,
  ps1 = 0.75,
  pp2 = 9/11,
  ps2 = 0.75,
  rho.N = 0,
  rho.P = 0,
  beta = 1
)
}
\arguments{
\item{N}{Number of classification instances.}

\item{s}{Number of positives.}

\item{pp1}{Precision of 1st classifier.}

\item{ps1}{Sensitivity of 1st classifier.}

\item{pp2}{Precision of 2nd classifier.}

\item{ps2}{Sensitivity of 2nd classifier.}

\item{rho.N}{Correlation of two classifiers' predictions
among negative cases.}

\item{rho.P}{Correlation of two classifiers' predictions
among positive cases.}

\item{beta}{beta.}
}
\description{
A function to show a distribution of the difference between two F beta scores.
}
\examples{
\dontrun{
fbscore <- F1.cond.two.b.rho(N=50, s=20,
                              pp1=9/11, ps1=0.75,
                              pp2=0.90, ps2=0.85,
                              rho.N=0, rho.P=0,
                              beta=1)
psF1::plotF1(fbscore, type=c("Density", "CDF")[2], xlim=c(-1,1))
}
}
