% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFbTwo.rho_v2.R
\name{powerFbTwo.rho}
\alias{powerFbTwo.rho}
\title{Power Calculation for F.beta Comparison Between Two Correlated Classifiers}
\usage{
powerFbTwo.rho(
  N = 100,
  s = 40,
  Fb1 = NULL,
  pp1 = 0.6,
  ps1 = 0.8,
  Fb2 = NULL,
  pp2 = 0.7,
  ps2 = 0.9,
  rho.N = 0,
  rho.P = 0,
  beta = 1,
  alternative = c("two.sided", "less", "greater"),
  normal.approx = c("auto", "TRUE", "FALSE"),
  alpha.level = 0.05
)
}
\arguments{
\item{N}{Number of classification instances.}

\item{s}{Number of positives.}

\item{Fb1}{F.beta score of 1st classifier.}

\item{pp1}{Precision of 1st classifier.}

\item{ps1}{Sensitivity of 1st classifier.}

\item{Fb2}{F.beta score of 2nd classifier.}

\item{pp2}{Precision of 2nd classifier.}

\item{ps2}{Sensitivity of 2nd classifier.}

\item{rho.N}{Correlation of two classifiers' predictions
among negative cases. rho.N will be estimated if rho.N = NULL.}

\item{rho.P}{Correlation of two classifiers' predictions
among positive cases. rho.P will be estimated if rho.P = NULL.}

\item{beta}{beta.}

\item{alternative}{A character string specifying the alternative hypothesis:
c("two.sided", "less", "greater").}

\item{normal.approx}{normal.approx = c("auto", "TRUE", "FALSE").}

\item{alpha.level}{alpha level.}
}
\description{
A function to calculate power for testing the difference in F.beta scores between two correlated classifiers.
}
\examples{
\dontrun{
powerFbTwo.rho(N=100, s=40,
                Fb1=NULL, pp1=0.6, ps1=0.8,
                Fb2=NULL, pp2=0.7, ps2=0.9,
                rho.N=0.2, rho.P=0.2,
                beta=1,
                alternative=c("two.sided", "less", "greater"),
                normal.approx=c("auto", "TRUE", "FALSE"),
                alpha.level=0.05)
}
}
