% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFbTwo.rho_v2.R
\name{testFbTwo.rho}
\alias{testFbTwo.rho}
\title{Test for Difference in F.beta Between Two Classifiers}
\usage{
testFbTwo.rho(
  data,
  rho.N = NULL,
  rho.P = NULL,
  Fb0 = "est",
  beta = 1,
  alternative = c("two.sided", "less", "greater"),
  normal.approx = c("auto", "TRUE", "FALSE")
)
}
\arguments{
\item{data}{A data frame (columns: actual outcomes,
C1 predicted outcomes, C2 predicted outcomes).
Encoded as 0 (negative) and 1 (positive).}

\item{rho.N}{Correlation of two classifiers' predictions
among negative cases. rho.N will be estimated if rho.N = NULL.}

\item{rho.P}{Correlation of two classifiers' predictions
among positive cases. rho.P will be estimated if rho.P = NULL.}

\item{Fb0}{Null distribution is generated using the estimated F.beta
if Fb0 = "est"; Null distribution is generated using all possible F.beta
if Fb0 = NULL.}

\item{beta}{beta.}

\item{alternative}{A character string specifying the alternative hypothesis:
c("two.sided", "less", "greater").}

\item{normal.approx}{normal.approx = c("auto", "TRUE", "FALSE").}
}
\description{
A function to test the difference in F.beta between two classifiers.
}
\examples{
\dontrun{
set.seed(2028)
data <- data_Generation(N=100, s=40,
                         pp1=0.6, ps1=0.8,
                         pp2=0.7, ps2=0.9,
                         rho.N=0.2, rho.P=0.2,
                         beta=1)
testFbTwo.rho(data)
testFbTwo.rho(data, rho.N=0, rho.P=0) # wider
}
}
